fix(prompts): guard version-history pagination, dedupe fetcher, extract orchestration hook
- Gate the deep-link, diff-menu, and scroll-sentinel pagination triggers on
  isFetching/isError (not just isFetchingNextPage) to stop a permanently
  failing page fetch from hammering the backend in an infinite retry loop.
- Reject version_type=mask when resolving the active version by id, both for
  rendering and for the playground-load fetch, since the by-id lookup isn't
  scoped the way the paginated list is.
- Fix the mobile version dropdown to highlight the resolved active version
  instead of the raw query param.
- Share one fetcher between usePromptVersionsById and
  usePromptVersionsByIdInfinite instead of duplicating the request logic.
- Reuse an already-loaded version from the paginated list instead of issuing
  a redundant by-id fetch when switching between loaded versions.
- Extract version-history orchestration (pagination, selection, active
  version resolution) out of PromptTab into usePromptVersionHistory.
